+++ gem/lib/pagy/toolbox/paginators/method.rb
@@ -15,13 +15,19 @@ class Pagy
   path = Pathname.new(__dir__)
   paginators.each { |symbol, name| autoload name, path.join(symbol.to_s) }
 
-  # Defines the pagy method. Include in the app controller/view.
+  # Pagy::Method defines the pagy method to be included in the app controller/view.
   Method = Module.new do
              protected
 
              define_method :pagy do |paginator = :offset, collection, **options|
-               merged_options = paginator == :calendar ? options : Pagy.options.merge(options)
-               Pagy.const_get(paginators[paginator]).paginate(self, collection, **merged_options)
+               options[:root_key] = 'page' if options[:jsonapi] # enforce 'page' root_key for JSON:API
+               options[:request]  = Request.new(options[:request] || request)
+               arguments          = if paginator == :calendar
+                                      [self, collection, options]
+                                    else
+                                      [collection, Pagy.options.merge(options)]
+                                    end
+               Pagy.const_get(paginators[paginator]).paginate(*arguments)
              end
            end
 end