+++ gem/lib/pagy/toolbox/paginators/offset.rb
@@ -4,16 +4,13 @@ class Pagy
   module OffsetPaginator
     module_function
 
-    # Return instance and page of results
-    def paginate(context, collection, **options)
-      context.instance_eval do
-        request = Request.new(options[:request] || self.request, options)
-        options[:page]  ||= request.resolve_page(options)
-        options[:limit]   = request.resolve_limit(options)
-        options[:count] ||= collection.instance_of?(Array) ? collection.size : OffsetPaginator.get_count(collection, options)
-        pagy = Offset.new(**options, request:)
-        [pagy, collection.instance_of?(Array) ? collection[pagy.offset, pagy.limit] : pagy.records(collection)]
-      end
+    # Return the Pagy::Offset instance and results
+    def paginate(collection, options)
+      options[:page]  ||= options[:request].resolve_page(options)
+      options[:limit]   = options[:request].resolve_limit(options)
+      options[:count] ||= collection.instance_of?(Array) ? collection.size : OffsetPaginator.get_count(collection, options)
+      pagy = Offset.new(**options)
+      [pagy, collection.instance_of?(Array) ? collection[pagy.offset, pagy.limit] : pagy.records(collection)]
     end
 
     # Get the collection count